+++ test/libsolidity/gasTests/abiv2_optimised.sol
@@ -17,9 +17,9 @@ contract C {
 // optimize-yul: true
 // ----
 // creation:
-//   codeDepositCost: 603000
-//   executionCost: 638
-//   totalCost: 603638
+//   codeDepositCost: 610600
+//   executionCost: 645
+//   totalCost: 611245
 // external:
 //   a(): 1029
 //   b(uint256): 2084