+++ Changelog.md
@@ -10,6 +10,7 @@ Compiler Features:
 Bugfixes:
  * Inline Assembly: Fix internal error when accessing incorrect constant variables.
  * Inheritance: Allow public state variables to override functions with dynamic memory types in their return values.
+ * JSON AST: Drop pointer suffix for memory reference types.
 
 
 ### 0.6.4 (2020-03-10)